% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassMode.R
\name{ClassMode}
\alias{ClassMode}
\title{Determing the mode from a vector of numbers.}
\usage{
ClassMode(classes)
}
\arguments{
\item{classes}{A charactor vector of elements to find the mode.}
}
\value{
The mode found.
}
\description{
\code{ClassMode} returns the most commonly occuring number from a vector of numbers.
}
\details{
\code{ClassMode} creates a frequency table based on a vector.  This table is then sorted by highest frequency
and returns the top element.  Note, in the case of a tie, the first element is chosen as opposed to a random tiebreaker or returning 
both elements.  For the sake of developing Random Forests of SDTs, this is a safer option.
}
