% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoftForestPredDepth1.R
\name{SoftForestPredDepth1}
\alias{SoftForestPredDepth1}
\alias{SoftForestPredDepth2}
\alias{SoftForestPredDepth3}
\alias{SoftForestPredDepth4}
\alias{SoftForestPredDepth5}
\alias{\code{SoftForestPredDepth}}
\alias{creates}
\alias{a}
\alias{single}
\alias{SDT}
\alias{for}
\alias{only}
\alias{classification}
\alias{to}
\alias{build}
\alias{the}
\alias{Random}
\alias{Forest}
\alias{of}
\alias{SDTs.}
\alias{This}
\alias{is}
\alias{equivalent}
\alias{tree}
\alias{in}
\alias{standard}
\alias{Forest.}
\alias{runs}
\alias{through}
\alias{\code{BestForestSplit}}
\alias{function}
\alias{each}
\alias{node}
\alias{given}
\alias{sets}
\alias{weights}
\alias{determined}
\alias{at}
\alias{previous}
\alias{steps}
\alias{order}
\alias{fit}
\alias{determine}
\alias{with}
\alias{prespecified}
\alias{depth.}
\alias{The}
\alias{final}
\alias{prediction}
\alias{returned}
\alias{weighted}
\alias{average}
\alias{terminal}
\alias{nodes}
\alias{probability}
\alias{that}
\alias{observation}
\alias{coded}
\alias{\code{1}}
\alias{which}
\alias{would}
\alias{be}
\alias{current}
\alias{class.}
\title{Building a single level for the Random Forest of SDTs.}
\usage{
SoftForestPredDepth1(trainresponse, train, test, num.features, ntry,
  keep = FALSE)
}
\arguments{
\item{trainresponse}{A vector of responses \code{0} and \code{1} for the training set with length equal to the number of observations in the training set.}

\item{train}{A matrix or data frame consisting of all possible variables to attempt for the training set.}

\item{test}{A matrix or data frame consisting of all possible variables to attempt for the test set.}

\item{num.features}{The number of variables in the dataset to possibly try.  The leftmost number of variables in the dataset are the variables chosen.}

\item{ntry}{The number of variables from the \code{num.features} to attempt to split.  This is useful for building random forests.  For a standard tree, choose \code{ntry = num.features}.}

\item{keep}{Logical if weights from a single observation should be kept.  Keep FALSE if a Random Forest is to be built.}
}
\value{
A vector of the final fitted probabilities for this classification.
}
\description{
Building a single level for the Random Forest of SDTs.
}
